Worked on modals
